python_api/requirements.txt
Flask==2.1.2
werkzeug==2.1.2
mysql-connector-python==8.0.28
pika==1.3.1
